Add timetable validators and API endpoints for MDS/MIT programs
